+++ NEWS.md
@@ -4,3 +4,5 @@
   `eldoc-diffstat-mode`; there's also a globalized minor mode
   `global-eldoc-diffstat-mode` that (by default) will turn on diffstat output in
   all supported major modes.
+- `eldoc-diffstat-lines` can be used to truncate/pad output to a specific number
+  of lines.